refactor: move jwt presentation to Presentation
